+++ src/nimble_tk/__init__.py
@@ -3,5 +3,6 @@ from .common import *
 from .notebook import *
 from .analytics.pandas_utils import *
 from .analytics.string_utils import *
-from .concurrent.tasks import *
+from .tasks.schedulers import *
+from .tasks.concurrent import *
 from .linux import *
