+++ src/nimble_tk/common/general_utils.py
@@ -119,115 +119,6 @@ class IterWrapper(object):
         return None
 
 
-class FixedDelayTaskScheduler(object):
-    """
-    A class for scheduling a task to be executed at regular delays. 
-    This will start the delay countdown only after the current task completes.
-
-    # sample code
-    # define the function to be run at regular intervals
-    def code_to_run_repeatedly():
-        ...
-        ...
-
-    # schedule the task to run every 2 seconds
-    scheduler = FixedDelayTaskScheduler(interval=2, code_to_run_repeatedly, log_prefix='[Some][Log][Prefix]')
-
-    # stop the time if required
-    scheduler.stop()
-    """
-
-    def __init__(self, interval, function, log_prefix='', log=True, init_interval=None, max_interval=None, *args, **kwargs):
-        """
-        Initializes the scheduler with specified parameters.
-
-        :param interval: The initial interval in seconds between task executions.
-        :param function: The function to be executed.
-        :param log_prefix: Optional prefix for log messages.
-        :param log: Flag to enable or disable logging.
-        :param init_interval: Optional initial delay before the first task execution.
-        :param max_interval: Maximum time allowed for an execution; beyond this, a new execution starts.
-        :param args: Additional arguments to pass to the function.
-        :param kwargs: Additional keyword arguments to pass to the function.
-        """
-        self._timer = None
-        self.interval = interval
-        self.function = function
-        self.args = args
-        self.kwargs = kwargs
-
-        # internal state
-        self.is_running = False
-        self.running_since = datetime.datetime(1970, 1, 1)
-        self.is_shutdown = False
-        self.log_prefix = log_prefix
-        self.log = log
-        self.max_interval = max_interval
-        self.__set_timer(interval=init_interval)
-
-    def _run(self):
-        """
-        Internal method to execute the scheduled function.
-        Manages the state and logs before and after the function execution.
-        """
-        set_thread_log_prefix(self.log_prefix)
-
-        if self.should_start_next_run():
-            self.__set_timer()
-            try:
-                self.is_running = True
-                self.running_since = datetime.datetime.now()
-                stopwatch = StopWatch()
-                if self.log:
-                    log_info(f"Function Started")
-                self.function(*self.args, **self.kwargs)
-                if self.log:
-                    log_info(
-                        f"Function Done - took time {stopwatch.get_time()}")
-            except:
-                log_traceback(
-                    f"Error in calling the Function - took time {stopwatch.get_time()}")
-            finally:
-                self.is_running = False
-
-    def should_start_next_run(self):
-        """
-        Determines whether the next function execution should start.
-
-        :return: True if the next execution should start, False otherwise.
-        """
-        # Your existing logic here...
-
-    def update_interval(self, new_interval):
-        """
-        Updates the interval between task executions.
-
-        :param new_interval: The new interval in seconds.
-        """
-        log_info(f"Updating interval from {self.interval} to {new_interval}")
-        self.interval = new_interval
-
-    def __set_timer(self, interval=None):
-        """
-        Sets the timer for the next execution of the task.
-
-        :param interval: The interval after which the task should be executed. Defaults to the current interval.
-        """
-        if not interval:
-            interval = self.interval
-
-        if not self.is_shutdown:
-            self._timer = threading.Timer(interval, self._run)
-            self._timer.daemon=True
-            self._timer.start()
-
-    def shutdown(self):
-        """
-        Shuts down the scheduler, cancelling any pending executions.
-        """
-        self.is_shutdown = True
-        self._timer.cancel()
-
 # ------------------------------------------------
 
 class TempClass(object):
